Add a plugin to combine several profile stacks into a single one
This plugin is intended to run after several profile stacks have been
generated using the Phase Map plugin. These profile stacks need to contain
the raw profile data (in JSON format) in their metadata, as is the case
when the Phase Map has generated them (and this metadata gets saved and
restored when saving the profile stacks as .tif files, too).

For convenience, the median is already plotted, and a future version might
fit a quadratic function, too, possibly exporting the entire shebang in
JSON or CSV format for further analysis.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>